latex-renderer/requirements.txt
requests~=2.26.0
nbconvert~=6.3.0
pyTelegramBotAPI~=4.2.0
pyyaml~=6.0
